move DummyCompiler into tests.utils as per Mark's suggestion
